+++ src/lib/core/commands/Command.ts
@@ -1,0 +1,7 @@
+
+export type Command = {
+    readonly type: unknown,
+    readonly data: unknown,
+}
+
+
