+++ src/lib/core/index.ts
@@ -2,3 +2,5 @@ export * from './Plugin';
 export * from './Property';
 export * from './EditorObject';
 export * from './FabricContext';
+export * from './commands';
+export * from './persistence';
