+++ src/lib/core/persistence/Persistance.ts
@@ -1,0 +1,6 @@
+
+export abstract class Persistance<T> {
+    abstract save(t: T): string;
+    abstract load(text: string): T;
+}
+
