+++ src/lib/core/persistence/index.ts
@@ -1,0 +1,2 @@
+export * from './Persistance';
+export * from './FabricPersistance';
